FF/requirements.txt
numpy==1.22.1
pandas==1.4.3
pyqlib==0.9.0
audtorch==0.6.4
torch==1.12.1
torch-geometric==2.1.0
pyyaml==6.0
scipy==1.9.0
scikit-learn==1.1.2
tqdm==4.64.0